+++ test/unit-tests/driver/cmakeFileApiDriver.test.ts
@@ -4,8 +4,8 @@ import { CMakeFileApiDriver } from '@cmt/drivers/cmakeFileApiDriver';
 import { CMakeDriver, CMakePreconditionProblemSolver } from '@cmt/drivers/cmakeDriver';
 import { CMakeGenerator, Kit } from '@cmt/kit';
 
-import { makeCodeModelDriverTestsuite } from './driver-codemodel-tests';
-import { makeDriverTestsuite } from './driver-test';
+import { makeCodeModelDriverTestsuite } from './CodeModel.test';
+import { makeDriverTestsuite } from './driver.test';
 
 async function cmakeFileApiDriverFactory(cmake: CMakeExecutable, config: ConfigurationReader, kit: Kit | null, workspaceFolder: string | null, preconditionHandler: CMakePreconditionProblemSolver, preferredGenerators: CMakeGenerator[]) {
     const d: CMakeDriver = await CMakeFileApiDriver.create(cmake, config, false, kit, null, null, null, workspaceFolder, preconditionHandler, preferredGenerators);
