+++ test/unit-tests/driver/cmakeServerDriver.test.ts
@@ -4,8 +4,8 @@ import * as cms_driver from '@cmt/drivers/cmakeServerDriver';
 import { CMakeDriver, CMakePreconditionProblemSolver } from '@cmt/drivers/cmakeDriver';
 import { CMakeGenerator, Kit } from '@cmt/kit';
 
-import { makeCodeModelDriverTestsuite } from './driver-codemodel-tests';
-import { makeDriverTestsuite } from './driver-test';
+import { makeCodeModelDriverTestsuite } from './CodeModel.test';
+import { makeDriverTestsuite } from './driver.test';
 
 async function cmakeServerDriverFactory(cmake: CMakeExecutable, config: ConfigurationReader, kit: Kit | null, workspaceFolder: string | null, preconditionHandler: CMakePreconditionProblemSolver, preferredGenerators: CMakeGenerator[]) {
     const d: CMakeDriver = await cms_driver.CMakeServerDriver.create(cmake, config, false, kit, null, null, null, workspaceFolder, preconditionHandler, preferredGenerators);
